Finished Combat
I have finished the combat within the game. I am yet to add in the win/lose condition for the combat as well as the exiting of the combat but the system works surprisingly well.

Next, I need to comment the code so I have any idea what is going on later.